OPERA-2318: skip tiles with missing baseline granules in historical mode.
